Fix remaining mettagrid tests (#2038)
Note: pufferlib currently installs `tests` package. So I had to move
some code to `metta.mettagrid.test_support` to avoid namespace clashes.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211018769476358)
