+++ mettagrid/tests/test_mettagrid.py
@@ -1,32 +1,9 @@
-from dataclasses import dataclass
 from typing import List, Tuple
 
 import numpy as np
 
 from metta.mettagrid.mettagrid_c import PackedCoordinate, dtype_actions
-from metta.mettagrid.test_support import EnvConfig
-
-
-# Constants from C++ code
-@dataclass
-class TokenTypes:
-    # Observation features
-    TYPE_ID_FEATURE: int = 0
-    GROUP: int = 1
-    HP: int = 2
-    FROZEN: int = 3
-    ORIENTATION: int = 4
-    COLOR: int = 5
-    CONVERTING_OR_COOLING_DOWN: int = 6
-    SWAPPABLE: int = 7
-    EPISODE_COMPLETION_PCT: int = 8
-    LAST_ACTION: int = 9
-    LAST_ACTION_ARG: int = 10
-    LAST_REWARD: int = 11
-    GLYPH: int = 12
-
-    # Object type IDs
-    WALL_TYPE_ID: int = 1
+from metta.mettagrid.test_support import EnvConfig, TokenTypes
 
 
 class ObservationHelper:
